Add metrics piechart
